third_party/blink/web_tests/external/wpt/editing/run/fontsize_1001-2000-expected.txt

This is a testharness.js-based test.
Found 42 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "foo<font size=4>[bar]</font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-size:large\\">bar</span>baz" but got "foo<font size=\\"4\\">bar</font>baz"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "foo<font size=+1>[bar]</font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-size:large\\">bar</span>baz" but got "foo<font size=\\"+1\\">bar</font>baz"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "foo<font size=+1>[bar]</font>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">bar</font>baz" but got "foo<font size=\\"+1\\">bar</font>baz"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<font size=4>foo<font size=1>b[a]r</font>baz</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font size=\\"4\\">foo<span style=\\"font-size:x-small\\">b</span>a<span style=\\"font-size:x-small\\">r</span>baz</font>" but got "<font size=\\"4\\">foo<font size=\\"1\\">b</font>a<font size=\\"1\\">r</font>baz</font>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "foo<span style=\\"font-size: xx-small\\">[bar]</span>baz" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "foo<span style=\\"font-size: xx-small\\">[bar]</span>baz" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<span style=\\"font-size: xx-small\\">foo[bar]baz</span>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<span style=\\"font-size: xx-small\\">foo[bar]baz</span>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["fontsize","4"]] "<span style=\\"font-size: large\\">foo[bar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-size:large\\">foobarbaz</span>" but got "<span style=\\"font-size:large\\">foo</span><font size=\\"4\\">bar</font><span style=\\"font-size:large\\">baz</span>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<span style=\\"font-size: large\\">foo<span style=\\"font-size: xx-small\\">b[a]r</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-size:large\\">foo<span style=\\"font-size:xx-small\\">b</span>a<span style=\\"font-size:xx-small\\">r</span>baz</span>" but got "<span style=\\"font-size:large\\">foo</span><span style=\\"font-size:xx-small\\">b</span><span style=\\"font-size:large\\">a</span><span style=\\"font-size:xx-small\\">r</span><span style=\\"font-size:large\\">baz</span>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<span style=\\"font-size: large\\">foo<span style=\\"font-size: xx-small\\">b[a]r</span>baz</span>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<span style=\\"font-size: large\\">foo<span style=\\"font-size: xx-small\\">b[a]r</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-size:large\\">foo<span style=\\"font-size:xx-small\\">b</span>a<span style=\\"font-size:xx-small\\">r</span>baz</span>" but got "<font size=\\"4\\">foo</font><span style=\\"font-size:xx-small\\">b</span><font size=\\"4\\">a</font><span style=\\"font-size:xx-small\\">r</span><font size=\\"4\\">baz</font>"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<span style=\\"font-size: large\\">foo<span style=\\"font-size: xx-small\\">b[a]r</span>baz</span>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:xx-small\\">foo<span style=\\"font-size:large\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"font-size:xx-small\\">foo</span><span style=\\"font-size:large\\">bar</span><span style=\\"font-size:xx-small\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:xx-small\\">foo<font size=\\"4\\">bar</font>baz</p>" but got "<p style=\\"\\"><span style=\\"font-size:xx-small\\">foo</span><font size=\\"4\\">bar</font><span style=\\"font-size:xx-small\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p style=\\"font-size: medium\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:medium\\">foo<span style=\\"font-size:large\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"font-size:medium\\">foo</span><span style=\\"font-size:large\\">bar</span>baz</p>"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p style=\\"font-size: medium\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:medium\\">foo<font size=\\"4\\">bar</font>baz</p>" but got "<p style=\\"\\"><span style=\\"font-size:medium\\">foo</span><font size=\\"4\\">bar</font>baz</p>"
[FAIL] [["fontsize","4"]] "<p style=\\"font-size: large\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:large\\">foobarbaz</p>" but got "<p style=\\"\\"><span style=\\"font-size:large\\">foo</span><font size=\\"4\\">barbaz</font></p>"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "<p style=\\"font-size: 2em\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:2em\\">foo<span style=\\"font-size:large\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"font-size:2em\\">foo</span><span style=\\"font-size:large\\">bar</span><span style=\\"font-size:2em\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "<p style=\\"font-size: 2em\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:2em\\">foo<font size=\\"4\\">bar</font>baz</p>" but got "<p style=\\"\\"><span style=\\"font-size:2em\\">foo</span><font size=\\"4\\">bar</font><span style=\\"font-size:2em\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["fontsize","3"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:xx-small\\">foo</span>bar<span style=\\"font-size:xx-small\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-size:xx-small\\">foo</span>bar<span style=\\"font-size:xx-small\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["fontsize","3"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["stylewithcss","false"],["fontsize","3"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:xx-small\\">foo</span>bar<span style=\\"font-size:xx-small\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-size:xx-small\\">foo</span>bar<span style=\\"font-size:xx-small\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontsize","3"]] "<p style=\\"font-size: xx-small\\">foo[bar]baz</p>" queryCommandValue("fontsize") before
  assert_equals: Wrong result returned expected "1" but got "0"
[FAIL] [["fontsize","3"]] "<p style=\\"font-size: medium\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-size:medium\\">foobarbaz</p>" but got "<p style=\\"\\"><span style=\\"font-size:medium\\">foo</span>barbaz</p>"
[FAIL] [["stylewithcss","true"],["fontsize","3"]] "<p style=\\"font-size: large\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:large\\">foo</span>bar<span style=\\"font-size:large\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-size:large\\">foo</span>bar<span style=\\"font-size:large\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontsize","3"]] "<p style=\\"font-size: large\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font>bar<font size=\\"4\\">baz</font></p>" but got "<p style=\\"\\"><span style=\\"font-size:large\\">foo</span>bar<font size=\\"4\\">baz</font></p>"
[FAIL] [["stylewithcss","true"],["fontsize","3"]] "<p style=\\"font-size: 2em\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:2em\\">foo</span>bar<span style=\\"font-size:2em\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-size:2em\\">foo</span>bar<span style=\\"font-size:2em\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["fontsize","3"]] "<p style=\\"font-size: 2em\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-size:2em\\">foo</span>bar<span style=\\"font-size:2em\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-size:2em\\">foo</span>bar<span style=\\"font-size:2em\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["fontsize","3"]] "foo<big>[bar]</big>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<big><span style=\\"font-size:medium\\">bar</span></big>baz" but got "foo<big style=\\"font-size:medium\\">bar</big>baz"
[FAIL] [["stylewithcss","true"],["fontsize","3"]] "foo<small>[bar]</small>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<small><span style=\\"font-size:medium\\">bar</span></small>baz" but got "foo<small style=\\"font-size:medium\\">bar</small>baz"
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "fo[o<font size=2>b]ar</font>baz" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontsize","4"]] "fo[o<font size=2>b]ar</font>baz" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontsize","4"]] "foo<font size=2>ba[r</font>b]az" queryCommandIndeterm("fontsize") before
  assert_equals: Wrong result returned expected true but got false
Harness: the test ran to completion.
